+++ src/core/operation.ts
